refactor: OAuth와 약관 동의 플로우 분리 및 home 기반 약관 리다이렉트 적용
